aistock-cronjob/requirements.txt
akshare
numpy
pandas
sqlalchemy
pymysql
requests
python-dotenv
redis
cryptography
openai
semhash
pypinyin
boto3